Using - separator in names + moving url-available to domain
